Skip tests if ./set-of-emails/obsoleted directory does not exist at Sisimai.dump test

+++ spec/obsoleted/sisimai_spec.rb
@@ -193,6 +193,7 @@ describe Sisimai do
 
     context 'valid JSON file' do
       samplejsons.each do |e|
+        break unless Dir.exist?('./set-of-emails/obsoleted')
         jf = File.open('./set-of-emails/obsoleted/' + e, 'r')
         js = jf.read
         jf.close
